+++ TissueLabeling/training/logging.py
@@ -6,7 +6,7 @@ import wandb
 from matplotlib.colors import BoundaryNorm, ListedColormap
 from PIL import Image
 
-from TissueLabeling.utils import crop, load_brains, mapping
+from TissueLabeling.brain_utils import crop, load_brains, mapping
 
 
 class Log_Images:
